Fix double loading of concern.
Can't load a concern twice. Otherwise it'll re-open the module and try
to add the included block multiple times. Which is not allowed and
results in an exception.

Basically, in developement, this meant that every time `to_prepare` was
run we would end up with an exception and need to restart the server.
Defeating the purpose of it entirely.