Pass args.last to layout when assuming a proc
The default layout is a proc that passes it's arguments down.
A nil error occurs if no args are passed to the proc. It's expected to
recieve the formats, which in the original method signature, is the last
argument.

actionview-5.0.7/lib/action_view/renderer/template_renderer.rb:75
actionview-5.0.7/lib/action_view/renderer/template_renderer.rb:386